# Optimize metrics query