--- a/packages/server/billing/helpers/adjustUserCount.ts
+++ b/packages/server/billing/helpers/adjustUserCount.ts
@@ -82,7 +82,8 @@ const addUser = async (orgIds: string[], userId: string) => {
       (oldOrganizationUser && oldOrganizationUser.newUserUntil) ||
       organization.periodEnd ||
       new Date()
-    return new OrganizationUser({orgId, userId, newUserUntil})
+    const tier = organization.tier
+    return new OrganizationUser({orgId, userId, newUserUntil, tier})
   })
   await r
     .table('OrganizationUser')