--- a/packages/server/database/migrations/20200701122639-addIndicesToUser.ts
+++ b/packages/server/database/migrations/20200701122639-addIndicesToUser.ts
@@ -1,4 +1,6 @@
-export const up = async function(r) {
+import {R} from 'rethinkdb-ts'
+
+export const up = async function(r: R) {
   try {
     await Promise.all([
       r