--- a/packages/server/database/migrations/20200702100758-addTierInactiveIndexOnOrgUser.ts
+++ b/packages/server/database/migrations/20200702100758-addTierInactiveIndexOnOrgUser.ts
@@ -1,0 +1,25 @@
+export const up = async function(r) {
+  try {
+    await Promise.all([
+      r
+        .table('OrganizationUser')
+        .indexCreate('tierInactive', [r.row('tier'), r.row('inactive')])
+        .run()
+    ])
+  } catch (e) {
+    console.log(e)
+  }
+}
+
+export const down = async function(r) {
+  try {
+    await Promise.all([
+      r
+        .table('OrganizationUser')
+        .indexDrop('tierInactive')
+        .run()
+    ])
+  } catch (e) {
+    console.log(e)
+  }
+}