--- a/packages/server/graphql/queries/suOrgCount.ts
+++ b/packages/server/graphql/queries/suOrgCount.ts
@@ -39,7 +39,7 @@ export default {
       return (r
         .table('OrganizationUser')
         .getAll(
-          ([tier, false] as unknown) as string,
+          ([tier, false] as unknown) as string, // super hacky type fix bc no fn overload is defined in the type file for this valid invocation
           ({index: 'tierInactive'} as unknown) as undefined
         )
         .filter({removedAt: null})