--- a/packages/server/graphql/queries/suOrgCount.ts
+++ b/packages/server/graphql/queries/suOrgCount.ts
@@ -1,4 +1,4 @@
-import {GraphQLInt, GraphQLBoolean, GraphQLString} from 'graphql'
+import {GraphQLInt} from 'graphql'
 import getRethink from '../../database/rethinkDriver'
 import {requireSU} from '../../utils/authorization'
 import {TierEnum as ETierEnum} from 'parabol-client/types/graphql'
@@ -7,16 +7,6 @@ import TierEnum from '../types/TierEnum'
 export default {
   type: GraphQLInt,
   args: {
-    ignoreEmailRegex: {
-      type: GraphQLString,
-      defaultValue: '',
-      description: "filter out users who's email matches this regular expression"
-    },
-    includeInactive: {
-      type: GraphQLBoolean,
-      defaultValue: false,
-      description: 'should organizations without active users be included?'
-    },
     minOrgSize: {
       type: GraphQLInt,
       defaultValue: 2,
@@ -28,45 +18,22 @@ export default {
       description: 'which tier of org shall we count?'
     }
   },
-  async resolve(_source, {ignoreEmailRegex, includeInactive, minOrgSize, tier}, {authToken}) {
+  // Only counts organizations with active users
+  async resolve(_source, {minOrgSize, tier}, {authToken}) {
     const r = await getRethink()
 
     // AUTH
     requireSU(authToken)
 
     // RESOLUTION
-    if (!ignoreEmailRegex && !includeInactive)
-      return (r
-        .table('OrganizationUser')
-        .getAll(
-          ([tier, false] as unknown) as string, // super hacky type fix bc no fn overload is defined in the type file for this valid invocation
-          ({index: 'tierInactive'} as unknown) as undefined
-        )
-        .filter({removedAt: null})
-        .group('orgId') as any)
-        .count()
-        .ungroup()
-        .filter((group) => group('reduction').ge(minOrgSize))
-        .count()
-        .run()
-
     return (r
       .table('OrganizationUser')
-      .filter((orgUser) => r.branch(includeInactive, true, orgUser('inactive').not()))
-      .filter({removedAt: null})
-      .filter((orgUser) => orgUser('tier').eq(tier))
-      .eqJoin('userId', r.table('User'))
-      .zip() as any)
-      .filter((result) =>
-        r.branch(
-          r(ignoreEmailRegex).eq(''),
-          true,
-          result('email')
-            .match(ignoreEmailRegex)
-            .eq(null)
-        )
+      .getAll(
+        ([tier, false] as unknown) as string, // super hacky type fix bc no fn overload is defined in the type file for this valid invocation
+        ({index: 'tierInactive'} as unknown) as undefined
       )
-      .group('orgId')
+      .filter({removedAt: null})
+      .group('orgId') as any)
       .count()
       .ungroup()
       .filter((group) => group('reduction').ge(minOrgSize))