--- a/packages/server/utils/setTierForOrgUsers.ts
+++ b/packages/server/utils/setTierForOrgUsers.ts
@@ -1,4 +1,5 @@
 import getRethink from '../database/rethinkDriver'
+import {TierEnum} from 'parabol-client/types/graphql'
 
 const setTierForOrgUsers = async (orgId: string) => {
   const r = await getRethink()
@@ -7,10 +8,10 @@ const setTierForOrgUsers = async (orgId: string) => {
     .getAll(orgId, {index: 'orgId'})
     .update(
       {
-        tier: r
+        tier: (r
           .table('Organization')
           .get(orgId)
-          .getField('tier')
+          .getField('tier') as unknown) as TierEnum
       },
       {nonAtomic: true}
     )