--- a/packages/server/utils/setTierForOrgUsers.ts
+++ b/packages/server/utils/setTierForOrgUsers.ts
@@ -1,3 +1,12 @@
+/*
+ * We only need to update on `OrganizationUser` objects
+ * with `removedAt = null`. Why? `OrganizationUser` row
+ * represents one instance of a "membership joining".
+ * Once `removedAt` field is populated, that `OrganizationUser`
+ * row will not be used again. In case the user is re-invited
+ * and rejoins the same org, a new `OrganizationUser` row
+ * will be created.
+ */
 import getRethink from '../database/rethinkDriver'
 import {TierEnum} from 'parabol-client/types/graphql'