Resolve #3952

## Acceptance Criteria
[x] tested that the following metrics query runs in well under 10s (the length of the timeout).
```
query getAllCounts ($OneWeekAgo: DateTime = "2016-01-01T00:00:00.000Z") {
  personalOrg: suOrgCount(includeInactive: false, tier: personal)
  proOrgs: suOrgCount(includeInactive: false, tier: pro)
  enterpriseOrgs: suOrgCount(includeInactive: false, tier: enterprise)
  personalUsers: suUserCount(includeInactive: false, tier: personal)
  proUsers: suUserCount(includeInactive: false, tier: pro)
  enterpriseUsers: suUserCount(includeInactive: false, tier: enterprise)

  logins: logins(after: $OneWeekAgo, isActive: false) {
  	total
  }
  
  signups: signups(after: $OneWeekAgo, isActive: false) {
  	total
  }
}
```

## Solution Discussion
Background: MR queries are auto parallelized in RethinkDB amongst shards and CPU cores. Since we currently have no shards, these queries are currently parallelized amongst CPU cores. However, there are certain operations that disable auto parallelization from occurring after. E.g. Joins and order-by. This is bc those commands require the results to be aggregated in one place. Any commands taking place after can only operate on the whole aggregated result set.

#### `suOrgCount` and `suUserCount`
- The original queries did an unnecessary join btwn `OrganizationUser` and `User`. The join is only necessary when we want to match against the user email address. Since metrics doesn't care about this, we can remove the join.
- The resulting query, after removing the join, finds all `OrganizationUser`s for each `Organization`. Even with the auto parallelizing done by RethinkDB, this query still completes on the order of a few seconds.
- Another idea is to use `DataLoader` to keep all the db records we want in memory on the app server. This would reduce our total fetching to `constant * N` => `N`. The downside of this approach is that we're moving CPU computation from DB to app server. Since we're forced to work with the cached objects, we lose the ability to use auto parallelization and convenient grouping API offered by RethinkDB. Imo, the upside didn't make up for the downside.
- The chosen solution adds a denormalized `tier` field to `OrganizationUser`.  Since our queries are limited to 1 table, this brings our query time down to the order of ms. Hooks were added to keep the value updated on tier upgrades and downgrades. 
  - We also added a compound index to `tier` and `inactive` fields on `OrganizationUser` called `tierInactive` index. Note that we cannot index `removedAt`, bc [secondary indexes in RethinkDB don't work on null values](https://rethinkdb.com/docs/secondary-indexes/javascript/). 

#### `logins` and `signups`
- We just add an index to `createdAt` and `lastSeenAt` fields on `User`. This also improves our query completion time from seconds to ms.

## Known Limitations
#### `suUserCount`
Currently the optimized query for `suUserCount` will return different results vs the unoptimized query. The reason is the latter ultimately counts `User`s, while the former counts `OrganizationUser` groupings. Currently, there are 100+ `OrganizationUser` objects with dangling references to `User`. I.e. `userId` is set but points to nothing in the `User `table. @mattkrick has been informed of the mismatch. The result sets will be identical once this mismatch is accounted for.