@tiffanyhan 2020-07-14
> This is an interesting problem because we're using our runtime DB for the purpose of analytics, so we've gotta be careful not to impact normal usage for the sake of a query that is literally run once a week.
> 
> With that in mind, here are my recommendations:
> 
> 1. I'd argue against adding the indexes on the `User` table.
>    `lastSeenAt` is updated every time the user opens the app. Indexing on such a hot field adds a penalty to normal app usage in exchange for a 700ms savings that is run 2x/day. I'd argue the penalty isn't worth it.
> 
> running a full-sized DB here are the results I got:
> 
> Without User indexes: 1600ms
> With User indexes: 980ms
> 
> 1. `tier` (or more accurately "highest tier achieved") is already denormalized onto the User object, so suUserCount can be simplified to `r.table('User').filter({tier}).count()`. this has the added benefit that a certain user won't be double counted. @jordanh can you confirm that for analytics it's better that a given user not be included in multiple tiers?
> 
> I think you made the right call denormalizing the tier onto the org user! it's a tough pill to swallow, always bums me out having to denormalize things, but it's 5x faster than anything I could come up with & that delta is only going to grow.
> For posterity, i found if i filtered out the inactives first, it could complete in about 1400ms, but yours completes in 300ms, so it's the clear winner.
> 
> ```
>   	r.db('actionDevelopment')
>          .table('OrganizationUser')
>       .getAll([tier, false], {index: 'tierInactive'}      )
>       .filter({removedAt: null})
>       .group('orgId')
>       .count()
>       .ungroup()
>       .filter((group) => group('reduction').ge(2))
>       .count()
> ```

Thanks for adding your review! 

I agree with you on the first point. The `logins` and `signups` queries aren't as problematic compared to `suOrgCount` and suUserCount` queries. I'm good with keeping those as is, in favor of faster app runtime for users.

On the second point, I actually had the same thought. The one thing about using `OrganizationUser` as the entry point for the `suUserQuery` is that we can verify `removedAt=null` then do the `group` and `ungroup` to dedupe by `userId`. I didn't want to assume that `OrganizationUser.removedAt=null` was equivalent to `User.inactive`. On testing the queries though, it looks like using `User` as the base for the query actually returns a lower number. I'm okay with using the suggested `User` query, if we don't need to qualify `OrganizationUser.removedAt = null`.