@tiffanyhan 2020-07-14
Per @mattkrick 's suggestions, changes have been made to: (1) remove indices from hot fields on `User` table; (2) Query the `User` table instead of `OrganizationUser` for `suUserCount`.