@mattkrick 2020-07-15
> (if a user is removed from an organization w/o any tier upgrades or downgrades occurring), but it seems good enough to me for metrics purposes.

that's a _really_ good find! this should also have a hook in `removeFromOrg`. I'll create an issue for it. Certainly doesn't block this PR. Great work!